requirements/requirements.txt
# data slice
numpy
pyyaml
rospkg
wind-pypcd
opencv-python

# gif_generator
imageio==2.21.1
rich

